Add the app's self dependency
When calling build like: `steal build cookbook/cookbook.js` the "appName" is cookbook.js, however this is not loaded. Rather cookbook.js is a dependency of the app. Due to the way JsDom works, the initial cookbook.js is not extracted during the fake dom process, so we need to add it during the dependency phase. Related to #131