Pass in dependencyMap to transpile
This updates pluginify to pass in a dependencyMap to transpile so that
it can correctly transpile modules that have been mapped or modules that
are not fully Normalized.  The dependencyMap is a mapping of
pre-normalized module names to the fully-normalized module names.

This is dependent on a change in transpile:
https://github.com/bitovi/transpile/pull/5