Pluginify: remove plugins prior to transpiling
This changes `makeOrderedTranspiledMinifiedGraph` to pluck out
dependencies that are not to be included in the build prior to doing
transpiling, ordering, and minifying.

Also changes pluginifier to pluck out all dependencies prior to
concating its sources. The result is that no plugins are included in the built JavaScript file.

Fixes #42