Upgrade UglifyJS
We need to use the tilde ~ to select the version of UglifyJS to use.
Previously we were using the exact version number which caused problems
because that version was no longer in npm. Using ~ will keep us up to
date on all patch releases without upgrading to the next minor release.