Fixes error reporting test
Fixes a test that reports an error when attempting to build a module
that isn't in the filesystem. Test was working fine but because Steal
does a console.log on errors it was sending a File Not Found message
into the console that could be misinterpreted. Fix is to swallow
console.logs during the running of the test and set it back afterwards.