Reset window.define before evaling in pluginify
Our System.define shim needs to have the same logic as our AMD shim.
That is, before executing code remove our `window.define` and replace it
with the original version, and then set back to our version after the
eval. This causes problems with "bad AMD". That is AMD that has it's
define statement at the bottom and therefore must be marked as a global.
Fixes #73