Set the value of `this` when evaling globals
For pluginify. This wraps the `eval()`ed source code with an anonymous function that sets the value of `this` to the window object. Because the functionis being called as `System.define` the value of `this` would naturally be `System`. This is problematic for types of globals that set the value of `this`, expecting that to be the window object. Fixes #112