import @config prior to main in production
This fixes issues surrounding loading the @config prior to the main in production. The issues were:

1. bundleSteal was not working with the old method.
2. System.defines that are dependencies of the @config would cause the app to never finish loading.

The new approach is to make the `@config` part of the main bundle in Steal and then to do:

```js
System.import("@config").then(function(){
  System.import("main");
});
```

Since `@config` is part of the main bundle it will cause the main bundle to be loaded, and then the config will be loaded prior to the main.

To fix this we had to make it so that other mains are not included in the bundles config. This is important for the multi-app build because the `@config` is part of each main's bundle. Without removing these extra bundles it would cause more than one app to be loaded.

Fixes #133