Add tests for building using bower plugin
This adds tests and a couple of minor fixes to make builds using the bower plugin work. There is no special logic for bower, but because the `@config` has dependencies that use a plugin, I had to move the trace up slightly so that it would capture the plugin dependency. Previously the trace didn't start until after `buildSteal.startup()`, which is too late.