Support export with module.exports
Babel spits out AMD code that makes use of an `exports` and `module` object.  This is different from other cjs->amd transpilation with uses `exports`, `module`, and `require`. Since Babel doesn't use `require` we need to shim this behavior.  It's as simple as checking the names of the arguments and that there is not an existing import for them. Fixes #189