Watch mode
The watch mode is a mode of the multi-build that continously builds as files in the dependency graph change. To use:

```shell
steal-tools --watch
```

This will use `package.json!npm` as the config (you can still pass `--config` and `--main` options if you want), do an intial build and watch for file changes which will trigger a rebuild.

Closes #80