Update to source-map dependency
This updates us to a patched version of our source-map dependency. This is a workaround for an issue with uglify where it keeps the original source map and adds new mappings, resulting in more generated lines than lines in the final source, which would previously throw an exception during concat. The fix is to ignore the extra line mappings.