live-reload: receive initial module
This updates live-reload to receive the initial module (the loader.main) when it first starts up. This is so that if you load a test page, or a demo page, or any page where the main is different from the application main, the server will be aware of the graph and live-reload will work.