Test against current version of Node and IO.js
This commit makes it so that we are testing against all of the important versions of Node/IO.js in Travis:

- Node 0.10.x
- Node 0.12.x
- IO.js 3.0.x

This works by calling the script `install-engine-dependencies` in Travis to install the appropriate version of Zombie. The correct version is defined in the package.json `engineDependencies` field.

If you are running tests locally you'll need to run `node_modules/.bin/install-engine-dependencies` yourself to get the correct version of Zombie.

Closes #292