Make work with envs configuration
envs is a special configuration added in Steal 0.11.0 that allows you to apply configuration in specific environments. To make this work in StealTools we need to change the env of the localSteal to `build-development`, this way apps can have special configuration that is applied to this environment (such as if you need the production build to happen a certain way). We only want to set this in the case of building to a multi-build target, because the envs configuration is a Steal-only feature. So we are passing in a new option, `localStealConfig` that is applied only to the localSteal.