Fixes race condition in live-reload server initialization.
This is related to https://github.com/donejs/donejs/issues/94

When can-serve and steal-tools live-reload are started at the same time with `donejs develop` it causes a race condition where if can-server loads before live-reload it will send a ping to the live-reload server to load the `System.main` into the stream. However it is possible that the graph hasn't been loaded by live-reload at this point, and if so an error will occur. The fix is to create a Promise out of the `graphStream` so that we don't write into it until it has been initiated.