Detect changes to dynamic bundles in graph/recycle
The done-autorender plugin and the can/view/stache/system plugin will both detect dynamic imports and add those to the `bundle` array on the loader. When recycling a module (such as in live-reload) we have an optimization that only reloads the dependency graph when there have been changes to the module's static dependencies. We need to compare the `bundle` array as well, and if that has changed also reload the dependency graph. Fixes #313