When merging graphs, keep all bundles for each node
When merging graphs we were using the new nodes for plugins. This meant that the old `bundles` property was being blown away so that plugins could wind up in a progressively loaded bundle and not in the main bundle they belonged to.  The fix is that when we use the new node we copy over the `bundles` onto that node when merging it back into the main graph. Fixes #298