Properly recycle virtual modules
This fixes #341

This is a good fix because it doesn't require any special configuration by plugins. Instead we detect any virtual modules that are created with System.define. When we diff a module to see if it has added/removed dependencies we run it through System.translate and System.instantiate. If translate produces virtual modules we then run the same process on the virtual modules. If anything has changed at all we know that a full trace is needed to get a fresh graph.