Reuse some config when loading bundles
This fixes #365. Some configuration might need to be reused when loading multiple bundles in multiple loaders. With the new progressive loading of package.jsons we need to keep the `npmContext` between loaders because each bundle will have a different set of NPM dependencies, and we need to build them **all** in the final output.

This will also require a fix in system-npm to reuse an existing npmContext.