Merge meta config in the Build Loader
For the bundle: false option we need to merge the meta configuration provided by `stealTools.build({ meta: {} })` with the config that will be in the user's configMain.

This changes it so that instead of normalizing meta in make_graph, instead there's a config hook in the trace that merge the meta. It's a little bit cleaner this way.