Add support for load.metadata.includedDeps
This adds support in live-reload for a new property:

```js
load.metadata.includedDeps = [
  "path/to/module"
];
```

This can be set by things like plugins, and this data is used to keep track of dependencies that are not part of the main dependency graph. live-reload will still track these files and emit for the parent module when the `includedDeps` changes.

In the client we now need to support receiving an array of module names. This is because includedDeps are not part of the real dependency graph, and therefore can trigger reloads in multiple parents.