Use exports to specify global exported values
This updates transform's `exports` property so that it can be used to specify values that should be exported.

For example if you have a module **foo** that looks like:

```js
module.exports = "foo bar";
```

And you set exports:

```js
exports: {
  "foo": "foo.bar"
}
```

Then it will set it so that `window.foo.bar === "foo bar"`.