Update tests to work with 1.0 module naming
This updates our tests to use the 1.0 module naming; when using the npm plugin the System.main will always include the package name. This means that when you add `main=` in the production script tag you must include the package name because the main bundle will be `dist/bundles/packageName/main`.