Include main bundle in other entry point bundles manifest
Since bundles correspond (roughly) to routes, the main bundle is always
going to be loaded when any other particular bundle is also loaded. So
the main bundle should be listed under each bundle's manifest entry.

This fixes #810. This is not a perfect fix, because for multi-build
projects, the main bundles won't be listed (nor should they be). This is
because the relationship between routes and bundles is not one-to-one,
		so we can't do this correctly without route information. #770
		seeks to solve that, and will result in the bundleManifest code
		being reconsidered.