Fix css bundles lookup in development bundles
Regular production bundles are always written out to "bundles/*", when a
custom "dest" is used, setting "bundlesPath" is needed to make sure the
defaults paths in System.paths point to the correct location of the bundles.

Development bundles are not written out to "bundles/*"; the fix in this case
consisted on adding "dest" to the bundle name used in the System.bundles
configuration added by steal-tools. This way the System bundle extension
can locate the file correctly when custom "dest" values are used.

Renaming the bundle name in the dependency graph brakes the logic to
write out the bundles, hence why the bundle name including "dest" is
only used in the bundle configuration output.

Closes #999