Run mocha programatically
Importing individual tests files in test/test would sometimes cause
the whole test suite not to run if any of the tests failed; also
sometimes the test suite finishes early for no apparent reason.

This approach is meant to fix these issues. Providing glob patterns
to the CLI is not practical in this case since the test folders holds
code that is not meant to be run my mocha directly.